+++ Source/XsollaStore/Public/XsollaStoreDataModel.h
@@ -7,6 +7,7 @@
 #include "XsollaStoreDefines.h"
 #include "XsollaUtilsDataModel.h"
 
+
 #include "XsollaStoreDataModel.generated.h"
 
 UENUM(BlueprintType)
@@ -130,6 +131,36 @@ public:
 };
 
 USTRUCT(BlueprintType)
+struct XSOLLASTORE_API FStoreBundleContent
+{
+	GENERATED_BODY()
+
+	UPROPERTY(BlueprintReadOnly, Category = "Bundle")
+	FString sku;
+
+	UPROPERTY(BlueprintReadOnly, Category = "Bundle")
+	FString name;
+
+	UPROPERTY(BlueprintReadOnly, Category = "Bundle")
+	FString type;
+
+	UPROPERTY(BlueprintReadOnly, Category = "Bundle")
+	FString description;
+
+	UPROPERTY(BlueprintReadOnly, Category = "Bundle")
+	FString image_url;
+
+	UPROPERTY(BlueprintReadOnly, Category = "Bundle")
+	int quantity;
+
+	UPROPERTY(BlueprintReadOnly, Category = "Bundle")
+	FStorePrice price;
+
+	UPROPERTY(BlueprintReadOnly, Category = "Bundle")
+	TArray<FVirtualCurrencyPrice> virtual_prices;
+};
+
+USTRUCT(BlueprintType)
 struct XSOLLASTORE_API FStoreItem
 {
 	GENERATED_BODY()
@@ -167,6 +198,18 @@ struct XSOLLASTORE_API FStoreItem
 	UPROPERTY(BlueprintReadOnly, Category = "Virtual Item")
 	FItemInventoryOptions inventory_options;
 
+	UPROPERTY(BlueprintReadOnly, Category = "Virtual Item Bundle")
+	FString bundle_type;
+
+	UPROPERTY(BlueprintReadOnly, Category = "Virtual Item Bundle")
+	FStorePrice total_content_price;
+
+	UPROPERTY(BlueprintReadOnly, Category = "Virtual Item Bundle")
+	TArray<FStoreBundleContent> content;
+
+	UPROPERTY(BlueprintReadOnly, Category = "Virtual Item Bundle")
+	TArray<FStoreItemAttribute> attributes;
+
 public:
 	FStoreItem()
 		: is_free(false){};
@@ -182,7 +225,15 @@ public:
 		, price(Item.price)
 		, virtual_prices(Item.virtual_prices)
 		, image_url(Item.image_url)
-		, inventory_options(Item.inventory_options){};
+		, inventory_options(Item.inventory_options)
+		, bundle_type(Item.bundle_type)
+		, total_content_price(Item.total_content_price)
+		, content(Item.content)
+		, attributes(Item.attributes)
+	{
+	}
+
+	FStoreItem(const struct FStoreBundle& Bundle);
 
 	bool operator==(const FStoreItem& Item) const
 	{
@@ -616,6 +667,60 @@ public:
 };
 
 USTRUCT(BlueprintType)
+struct XSOLLASTORE_API FStoreBundle
+{
+	GENERATED_BODY()
+
+	UPROPERTY(BlueprintReadOnly, Category = "Bundle")
+	FString sku;
+
+	UPROPERTY(BlueprintReadOnly, Category = "Bundle")
+	FString name;
+
+	UPROPERTY(BlueprintReadOnly, Category = "Bundle")
+	TArray<FStoreGroup> groups;
+
+	UPROPERTY(BlueprintReadOnly, Category = "Bundle")
+	TArray<FStoreItemAttribute> attributes;
+
+	UPROPERTY(BlueprintReadOnly, Category = "Bundle")
+	FString type;
+
+	UPROPERTY(BlueprintReadOnly, Category = "Bundle")
+	FString bundle_type;
+
+	UPROPERTY(BlueprintReadOnly, Category = "Bundle")
+	FString description;
+
+	UPROPERTY(BlueprintReadOnly, Category = "Bundle")
+	FString image_url;
+
+	UPROPERTY(BlueprintReadOnly, Category = "Bundle")
+	FString is_free;
+
+	UPROPERTY(BlueprintReadOnly, Category = "Bundle")
+	FStorePrice price;
+
+	UPROPERTY(BlueprintReadOnly, Category = "Bundle")
+	FStorePrice total_content_price;
+
+	UPROPERTY(BlueprintReadOnly, Category = "Bundle")
+	TArray<FVirtualCurrencyPrice> virtual_prices;
+
+	UPROPERTY(BlueprintReadOnly, Category = "Bundle")
+	TArray<FStoreBundleContent> content;
+};
+
+USTRUCT(BlueprintType)
+struct XSOLLASTORE_API FStoreListOfBundles
+{
+	GENERATED_BODY()
+
+	UPROPERTY(BlueprintReadOnly, Category = "Bundle")
+	TArray<FStoreBundle> items;
+};
+
+USTRUCT(BlueprintType)
 struct XSOLLASTORE_API FStoreUnitItem
 {
 public:
@@ -666,7 +771,7 @@ public:
 };
 
 USTRUCT(BlueprintType)
-struct XSOLLASTORE_API FStoreCoupoBonusItem
+struct XSOLLASTORE_API FStoreCouponBonusItem
 {
 public:
 	GENERATED_BODY()
@@ -678,7 +783,7 @@ public:
 	int32 quantity;
 
 public:
-	FStoreCoupoBonusItem()
+	FStoreCouponBonusItem()
 		: quantity(0){};
 };
 
@@ -689,7 +794,7 @@ public:
 	GENERATED_BODY()
 
 	UPROPERTY(BlueprintReadOnly, Category = "Coupon Data")
-	TArray<FStoreCoupoBonusItem> bonus;
+	TArray<FStoreCouponBonusItem> bonus;
 
 	UPROPERTY(BlueprintReadOnly, Category = "Coupon Data")
 	bool is_selectable;
@@ -761,6 +866,32 @@ public:
 };
 
 USTRUCT(BlueprintType)
+struct XSOLLASTORE_API FStoreDiscount
+{
+	GENERATED_BODY()
+
+	UPROPERTY(BlueprintReadOnly, Category = "Discount")
+	FString percent;
+};
+
+USTRUCT(BlueprintType)
+struct XSOLLASTORE_API FStorePromocodeRewardData
+{
+	GENERATED_BODY()
+
+	UPROPERTY(BlueprintReadOnly, Category = "Promocode Reward Data")
+	TArray<FStoreCouponBonusItem> bonus;
+
+	/** Percent discount. The price of cat will be decreased using a value calculated by using this percent and then rounded to 2 decimal places. */
+	UPROPERTY(BlueprintReadOnly, Category = "Promocode Reward Data")
+	FStoreDiscount discount;
+
+	/** If 'true', the user should choose the bonus before redeeming a promo code. */
+	UPROPERTY(BlueprintReadOnly, Category = "Promocode Reward Data")
+	bool is_selectable;
+};
+
+USTRUCT(BlueprintType)
 struct FXsollaJsonVariant
 {
 	GENERATED_BODY()
@@ -826,3 +957,20 @@ enum class EXsollaVariantTypes : int8
 	NetworkGUID = 38,
 	Custom = 0x40
 };
+
+inline FStoreItem::FStoreItem(const struct FStoreBundle& Bundle)
+{
+	this->description = Bundle.description;
+	this->groups = Bundle.groups;
+	this->image_url = Bundle.image_url;
+	this->is_free = Bundle.is_free == "true";
+	this->name = Bundle.name;
+	this->price = Bundle.price;
+	this->sku = Bundle.sku;
+	this->type = Bundle.type;
+	this->virtual_prices = Bundle.virtual_prices;
+	this->bundle_type = Bundle.bundle_type;
+	this->total_content_price = Bundle.total_content_price;
+	this->content = Bundle.content;
+	this->attributes = Bundle.attributes;
+}